Renamed 'passwords' route to 'content'
Along with associated modules/classes.
