+++ .gitignore
@@ -27,7 +27,7 @@
 /connect.lock
 /coverage/*
 /libpeerconnection.log
-npm-debug.log
+*debug.log
 testem.log
 /typings
