+++ src/app/content/components/category/category.component.scss
