+++ src/app/content/components/category/category.component.ts
