+++ src/app/content/components/item/item.component.scss
