+++ src/app/content/content.component.scss
