+++ src/app/content/models/index.ts
