+++ src/app/effects/items.ts
@@ -9,9 +9,9 @@ import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
 import { combineLatest } from 'rxjs/observable/combineLatest';
 
-import { BackendService } from '../passwords/services/backend.service';
+import { BackendService } from '../content/services/backend.service';
 import * as item from '../actions/items';
-import { ICategory } from '../passwords/models';
+import { ICategory } from '../content/models';
 
 @Injectable()
 export class ItemEffects {
